Pre-Competition Update 1
Here, we set autonomous to go forward for only 4 seconds instead of 10. We also added the tower to the motor configuration and user control. We changed all our while loops in the user control to if statements to allow the use of the lift or chain while driving. Last thing we did was only send 65% power to the wheels to reduce overheating.